Update control.lua
Add local plates_to_refund variable to update_recipes function to track how many alien-plates are currently being consumed (I'm not sure if the current input_inventory is deleted during an update, if so you can retrieve the current inventory amount and add into this total). 
Retrieve entity's inventory and do an insert of the quantity that was deleted.