Update control.lua
Modifies the core mechanism for updating, keeps most of the other logic identical. It scans the surface in batches using luaChunkIterator. These iterators are saved in the global table. 
Removed reference to on_load (this is known to cause desyncs-- on init will run on a previous save if this is the first time running with this mod enabled). 
Added some debugging options. When debug_mode is set to true, you will see a red circle appear for 30 frames near where was scanned (it was easier syntax than the rectangle, but it should cover a larger rectangle. 
I have done some testing but I recommend additional testing be performed.